最近检查了下代码，BUG什么的倒是没找到，变量类型和变量命名以及函数命名有点乱，决定用cstdint和驼峰命名整个替换一下

+++ Engine/Inc/Gear/Input/Event.h
@@ -5,6 +5,7 @@
 
 #include<map>
 #include<functional>
+#include<cstdint>
 
 class Event
 {
@@ -12,17 +13,17 @@ public:
 
 	Event();
 
-	void operator-=(const int id);
+	void operator-=(const uint64_t id);
 
- 	int operator+=(const std::function<void(void)>& func);
+	uint64_t operator+=(const std::function<void(void)>& func);
 
 	void operator()();
 
 private:
 
-	int idx;
+	uint64_t idx;
 
-	std::map<int, std::function<void(void)>> functions;
+	std::map<uint64_t, std::function<void(void)>> functions;
 
 };
